refactor: move face++ api fetching logic to face++ api client
