--- backend/src/clients/facepluplus_api_client.py
@@ -1,4 +1,0 @@
-from tornado.httpclient import AsyncHTTPClient
-
-class FacePlusPlusApiClient:
-    pass
