+++ backend/src/repository/__init__.py
@@ -4,3 +4,4 @@ from __future__ import absolute_import, division, print_function
 
 from .faces_repository import FacesRepository
 from .photos_repository import PhotosRepository
+from .emotions_repository import EmotionsRepository
